Separate core from main
