+++ dev_plans/20250528_core_refactor.md
@@ -27,53 +27,53 @@
 *   [x] **0.22. `meal_planner/__init__.py`**: Add module docstring (if file exists, or create and add) explaining the purpose of the main `meal_planner` package.
 
 **Phase 1: Create and Populate `meal_planner/core.py`**
-*   [ ] **1.1. Create File:** Create `meal_planner/core.py` (and add its module docstring).
-*   [ ] **1.2. Add Imports:** Add necessary base imports to `core.py` (e.g., `logging`, `Path` from `pathlib`, `httpx`, `FastAPI` from `fastapi`, `FastHTMLWithLiveReload`, `Theme` from `fasthtml.common`, `ASGITransport` from `httpx`, `StaticFiles` from `starlette.staticfiles`).
-*   [ ] **1.3. Add API Router Import:** Add `from meal_planner.api.recipes import API_ROUTER as RECIPES_API_ROUTER`.
-*   [ ] **1.4. Initialize Logger:** Add `logger = logging.getLogger(__name__)`.
-*   [ ] **1.5. Define `STATIC_DIR`:**
+*   [x] **1.1. Create File:** Create `meal_planner/core.py` (and add its module docstring).
+*   [x] **1.2. Add Imports:** Add necessary base imports to `core.py` (e.g., `logging`, `Path` from `pathlib`, `httpx`, `FastAPI` from `fastapi`, `FastHTMLWithLiveReload`, `Theme` from `fasthtml.common`, `ASGITransport` from `httpx`, `StaticFiles` from `starlette.staticfiles`).
+*   [x] **1.3. Add API Router Import:** Add `from meal_planner.api.recipes import API_ROUTER as RECIPES_API_ROUTER`.
+*   [x] **1.4. Initialize Logger:** Add `logger = logging.getLogger(__name__)`.
+*   [x] **1.5. Define `STATIC_DIR`:**
     *   Move/copy `STATIC_DIR = Path(__file__).resolve().parent.parent / "static"` (adjusting path for `core.py`'s location).
-*   [ ] **1.6. Define `app`:**
+*   [x] **1.6. Define `app`:**
     *   Move `app = FastHTMLWithLiveReload(hdrs=(Theme.blue.headers()))` from `main.py` to `core.py`.
-*   [ ] **1.7. Define `rt`:**
+*   [x] **1.7. Define `rt`:**
     *   Move `rt = app.route` from `main.py` to `core.py`.
-*   [ ] **1.8. Define `api_app`:**
+*   [x] **1.8. Define `api_app`:**
     *   Move `api_app = FastAPI()` from `main.py` to `core.py`.
     *   Move `api_app.include_router(RECIPES_API_ROUTER)` to `core.py`.
-*   [ ] **1.9. Define `internal_client`:**
+*   [x] **1.9. Define `internal_client`:**
     *   Move `internal_client = httpx.AsyncClient(transport=ASGITransport(app=app), ...)` from `main.py` to `core.py`, ensuring `app` refers to `core.app`.
-*   [ ] **1.10. Define `internal_api_client`:**
+*   [x] **1.10. Define `internal_api_client`:**
     *   Move `internal_api_client = httpx.AsyncClient(transport=ASGITransport(app=api_app), ...)` from `main.py` to `core.py`, ensuring `api_app` refers to `core.api_app`.
 
 **Phase 2: Refactor `meal_planner/main.py`**
-*   [ ] **2.1. Remove Definitions:** Delete the definitions of `STATIC_DIR`, `app`, `rt`, `api_app` (and its `include_router`), `internal_client`, and `internal_api_client` from `main.py`.
-*   [ ] **2.2. Add Core Imports:** Add `from meal_planner.core import app, rt, api_app, internal_client, internal_api_client, STATIC_DIR` to `main.py`.
-*   [ ] **2.3. Verify Mounts:** Ensure `app.mount("/static", StaticFiles(directory=STATIC_DIR), name="static")` and `app.mount("/api", api_app)` in `main.py` correctly use the imported `app`, `STATIC_DIR`, and `api_app` from `core`.
-*   [ ] **2.4. Logging Configuration:** Keep `logging.basicConfig(...)` in `main.py` as it's the application entry point. The `logger = logging.getLogger(__name__)` can also remain.
+*   [x] **2.1. Remove Definitions:** Delete the definitions of `STATIC_DIR`, `app`, `rt`, `api_app` (and its `include_router`), `internal_client`, and `internal_api_client` from `main.py`.
+*   [x] **2.2. Add Core Imports:** Add `from meal_planner.core import app, rt, api_app, internal_client, internal_api_client, STATIC_DIR` to `main.py`.
+*   [x] **2.3. Verify Mounts:** Ensure `app.mount("/static", StaticFiles(directory=STATIC_DIR), name="static")` and `app.mount("/api", api_app)` in `main.py` correctly use the imported `app`, `STATIC_DIR`, and `api_app` from `core`.
+*   [x] **2.4. Logging Configuration:** Keep `logging.basicConfig(...)` in `main.py` as it's the application entry point. The `logger = logging.getLogger(__name__)` can also remain.
 
 **Phase 3: Update Router Modules (`pages.py`, `actions.py`, `ui_fragments.py`)**
-*   [ ] **3.1. `meal_planner/routers/pages.py`:**
+*   [x] **3.1. `meal_planner/routers/pages.py`:**
     *   Change `from meal_planner.main import rt, internal_client, internal_api_client`
         TO `from meal_planner.core import rt, internal_client, internal_api_client`.
-*   [ ] **3.2. `meal_planner/routers/actions.py`:**
+*   [x] **3.2. `meal_planner/routers/actions.py`:**
     *   Change `from meal_planner.main import internal_client, rt`
         TO `from meal_planner.core import internal_client, rt`.
-*   [ ] **3.3. `meal_planner/routers/ui_fragments.py`:**
+*   [x] **3.3. `meal_planner/routers/ui_fragments.py`:**
     *   Change `from meal_planner.main import rt`
         TO `from meal_planner.core import rt`.
 
 **Phase 4: Update Test Files**
-*   [ ] **4.1. Identify Patch Targets:** Systematically search all test files for `patch("meal_planner.main.app")`, `patch("meal_planner.main.rt")`, `patch("meal_planner.main.internal_client")`, `patch("meal_planner.main.internal_api_client")`, and any direct imports of these from `meal_planner.main`.
-*   [ ] **4.2. Update Patch Targets:** Modify identified patch targets and direct imports to point to `meal_planner.core` (e.g., `patch("meal_planner.core.internal_client")`).
+*   [x] **4.1. Identify Patch Targets:** Systematically search all test files for `patch("meal_planner.main.app")`, `patch("meal_planner.main.rt")`, `patch("meal_planner.main.internal_client")`, `patch("meal_planner.main.internal_api_client")`, and any direct imports of these from `meal_planner.main`.
+*   [x] **4.2. Update Patch Targets:** Modify identified patch targets and direct imports to point to `meal_planner.core` (e.g., `patch("meal_planner.core.internal_client")`).
 
 **Phase 5: Verification and Code Quality**
-*   [ ] **5.1. Format Code:** Run `uv run ruff format .`.
-*   [ ] **5.2. Lint Code:** Run `uv run ruff check --fix .`.
-*   [ ] **5.3. Run Tests:** Execute `./run_fast_coverage.sh`.
-*   [ ] **5.4. Resolve Test Issues:** Debug and fix any test failures, ensuring all tests pass.
-*   [ ] **5.5. Verify Coverage:** Confirm test coverage remains at 100%.
-*   [ ] **5.6. Manual Test (Optional but Recommended):** Run the application locally (`uv run modal serve deploy.py`) and perform a quick functional check of key features.
+*   [x] **5.1. Format Code:** Run `uv run ruff format .`.
+*   [x] **5.2. Lint Code:** Run `uv run ruff check --fix .`.
+*   [x] **5.3. Run Tests:** Execute `./run_fast_coverage.sh`.
+*   [x] **5.4. Resolve Test Issues:** Debug and fix any test failures, ensuring all tests pass.
+*   [x] **5.5. Verify Coverage:** Confirm test coverage remains at 100%.
+*   [x] **5.6. Manual Test (Optional but Recommended):** Run the application locally (`uv run modal serve deploy.py`) and perform a quick functional check of key features. (Skipped)
 
 **Phase 6: Commit**
-*   [ ] **6.1. Suggest Commit Message.**
-*   [ ] **6.2. User Confirmation and Commit.**
+*   [x] **6.1. Suggest Commit Message.**
+*   [x] **6.2. User Confirmation and Commit.**
