+++ meal_planner/form_processing.py
@@ -3,7 +3,7 @@
 from starlette.datastructures import FormData
 
 
-def _parse_recipe_form_data(form_data: FormData, prefix: str = "") -> dict:
+def parse_recipe_form_data(form_data: FormData, prefix: str = "") -> dict:
     """Parses recipe form data into a dictionary, handling optional prefix."""
     name_value = form_data.get(f"{prefix}name")
     name = name_value if isinstance(name_value, str) else ""
